fixed nextjs app
